[tuneup.c] added cast from time_t to long, after warning under HP-UX: tuneup.c:1131: warning: format '%ld' expects type 'long int', but argument 3 has type 'time_t' tuneup.c:1133: warning: format '%ld' expects type 'long int', but argument 2 has type 'time_t' (I don't know if it is because time_t is unsigned, or has a different width)
[From SVN r7238 (trunk)]

+++ tune/tuneup.c
@@ -1128,9 +1128,9 @@ all (const char *filename)
   /* End of tuning */
   time (&end_time);
   fprintf (f, "/* Tuneup completed successfully, took %ld seconds */\n",
-           end_time - start_time);
+           (long) (end_time - start_time));
   if (verbose)
-    printf ("Complete (took %ld seconds).\n", end_time - start_time);
+    printf ("Complete (took %ld seconds).\n", (long) (end_time - start_time));
 
   fclose (f);
 }
